Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -44,5 +44,5 @@ def get_average_age_for_season(data, season)
   # code here
   age_sum = 0
   data[season].find do |contestant|
-    age_sum += contestant["age"].to_i
+    age_sum += contestant["age"].to_f
 end
